requirements.txt
box-ai-agents-toolkit>=0.1.3
box-sdk-gen>=1.17.0
python-dotenv>=1.2.1
colorlog>=6.10.1
boto3>=1.34.0
python-docx>=1.1.0
reportlab>=4.0.0
python-pptx>=0.6.21

